requirements.txt
click
PyYAML